Move Relic Data to separate file
